Reorganized tests dir to make it concise
