+++ server/tests/e2e/queries/Friend.ts
