+++ server/tests/e2e/resolvers/User/mutation/signInEmail.test.ts
@@ -1,6 +1,7 @@
+import { signInEmailMutation, signUpMutation } from '../../../queries';
+
 import { request } from 'graphql-request';
-import { signInEmailMutation, signUpMutation } from '../../../setup/queries';
-import { testHost } from '../../../setup/testSetup';
+import { testHost } from '../../../testSetup';
 
 describe('signInEmail', () => {
   it('should throw error when user does not exists', async () => {
