+++ server/tests/e2e/resolvers/User/mutation/signInWithFacebook.test.ts
@@ -1,8 +1,8 @@
 import * as AuthUtils from '../../../../../src/utils/auth';
 
 import { request } from 'graphql-request';
-import { signInWithFacebook } from '../../../setup/queries';
-import { testHost } from '../../../setup/testSetup';
+import { signInWithFacebook } from '../../../queries';
+import { testHost } from '../../../testSetup';
 
 describe('signInWithFacebook', () => {
   it('should signIn user wigh Facebook', async () => {
