+++ server/tests/e2e/resolvers/upload.test.ts
@@ -2,7 +2,7 @@ import * as azureUtils from '../../../src/utils/azure';
 
 import FormData from 'form-data';
 import fetch from 'node-fetch';
-import { testHost } from '../setup/testSetup';
+import { testHost } from '../testSetup';
 
 describe('Resolver - File', () => {
   beforeAll(async () => {
