+++ server/tests/e2e/testSetup.ts
@@ -6,10 +6,10 @@ import NodeWebSocket from 'ws';
 import { PrismaClient } from '@prisma/client';
 import { SubscriptionClient } from 'subscriptions-transport-ws';
 import { WebSocketLink } from 'apollo-link-ws';
-import { createApp } from '../../../src/app';
+import { createApp } from '../../src/app';
 import { exec } from 'child_process';
 import express from 'express';
-import { startServer } from '../../../src/server';
+import { startServer } from '../../src/server';
 
 // @ts-ignore
 global.Headers = global.Headers || Headers;
